Snapshot
Adds a client C file to call the generated library and a helper .c and .h file
for the generated library to use to demonstrate `extern`.  To the primer itself,
adds information about exporting functions and extern procs and globals.  Set
up is there for extern records and typedefs, but not fleshed out and there are
quite a few TODOs that I may or may not get to.